apps/media: add thread annotations and thread checker
This CL adds thread annotations to all files that use mutexes and
replaces a private thread_checker.h with ftl's. std::mutex was used
in some files - it's been replaced by ftl::Mutex.

There are still many places in the media code where thread_checker
should be used. This CL just addresses the instances in which the
private thread_checker was used.

Change-Id: I38ac2aa408a726a9f2de7c8e5ddcae075ee237a5