apps/media: add audio volume types
This CL adds two types used to express audio volume:
1) Gain expresses relative volume in logarithmic decibels.
2) Level is a template for linear volume.
Gain is a high-level notion of volume. It's well-known in the audio
discipline, but has performance issues where signal processing is
concerned. Level is defined for practical signal processing purposes.
It's a template, because the underlying representation may vary
depending on the types of samples we're processing. As of this CL,
the only specialization implemented for Level uses float as an
underlying type. This is useful when processing float samples. In
future CLs, we'll introduce at least one fixed-point representation
based on an unsigned integer type.

Change-Id: If698f7c6123d6328c34abb22b687da533fa84274