apps/media: add MixdownTable type
A MixdownTable specifies mix levels for each input/output channel
pair on a mixer input. It's basically just an MxN array of Level
values where M is the number of input channels and N is the number
of output channels. MixdownTables are mutable for convenience in
creating new tables. Once submitted to the mixer, however, they
won't be mutated. The creation static methods, therefore, return
std::unique_ptr<MixdownTable>.

We'll probably have a similar type that's an array of Gain values.
That will come in a later CL.

Change-Id: I50bca30418ba2e24fb2722b547bcee92659a4253