apps/media: add mixer
This CL adds the mixer and the abstract definition of a mixer input.
In addition, a README.md is supplied explaining how the mixer works.
The code in this CL represents only a small part of the functioning
mixer - most of the code will be submitted in the form of a concrete
MixerInput implementation. There are no unit tests for the mixer
itself. Instead, Mixer and MixerInputImpl are tested together. The
relevant tests will appear in a subsequent CL.

Change-Id: I36ece784a3063228b8163fe932c156a4a47d6d71