apps/media: add mixer input implementation
This is the bulk of the core mixer code in the form of a mixer input
implementation and unit tests for the mixer. The README.md file
submitted previously describes how it works. This CL includes
template specializations for floats only. Future CLs will address
various integer sample formats and fixed-point levels.

Change-Id: I797b5367e15b7df7c1cf74d9fc7299276ee82f2c