apps/media: add file source
This CL adds a file source to the items that media_service can be
asked to create. A file source allows read access to a file via the
SeekingReader interface, which demuxes consume. There are no tests
at the moment. I'm still figuring out how to implement multiple tests
as one ApplicationImplBase.

Change-Id: I5be741b5b01481eecaa312321c582875a4fda25d