apps/media: allow inexact pts rate calculations
Packets contain a 'pts rate' in the form of a ratio of two uint32s.
This ratio indicates the units of the pts field (ticks per second).
We often want to change the pts rate of a packet so the pts has the
desired units. In theory, the formula we want is:

    new_pts = old_pts * new_pts_rate / old_pts_rate

The pts values themselves are int64s.

We're currently doing this by first calculating
new_pts_rate / old_pts_rate as another ratio of two uint32s. In some
cases, this value can't be represented exactly as a ratio of uint32s.
Prior to this CL, that imprecision caused a DCHECK to fire.

This CL addresses that issue by performing the division in a way that
tolerates imprecision. This means that, in some cases, the resulting
new_pts won't be exactly the intended value, but the consequences of
this are negligible. TODOs have been added to do this differently so
that exact results are always produced. This will probably involve a
128-bit muldiv.

Change-Id: I3f6d5a87bdce340f7158e4228725bd2bcda91ab2