apps/media: add preliminary usb audio support
We'd like to be able to use the existing usb audio output capability
demonstrated in magenta's wavplay app. This isn't 'real' driver
support in the sense of implementing the HAL we want, but it has the
potential to enable some development and demos until that's ready.
This CL adds usb audio support to audio_server.

Change-Id: I568c579784d076797f9f8f2b42c73897f08d84b9