apps/media: package video rendering as a fidl component
Video renderering has previously been offered in the form of
objects that need to run in-proc in the application. This CL packages
video rendering as a fidl component vended by MediaService. The
app (video_player, in this case) hosts the view. The code that
implements the bulk of the video renderer has been moved from
apps/media/lib to apps/media/src/video, and video_render_impl.h/.cc
has been added to apps/media/src/media_service to implement the
video-rendering ViewProvider.

Change-Id: I218498367274416159a69cac7a5349f7255647be