apps/media: add keyframe flag to packets
Packets produced by the ffmpeg demux and consumed by ffmpeg decoders
are marked with a keyframe flag indicating whether the packet
contains a keyframe. This flag was being dropped, because framework
and fidl packets don't have this flag. This CL adds the flag to both
packet implementations. Note that the messages ffmpeg generates
complaining about missing keyframe flags are coming from the demux
as it parses the ogg container and are not related to the absence of
keyframe flags on packets.

Change-Id: I3efc7e6a11ed2678ca7132fd2ceec42c92c76936