apps/media: add audio capturer to media services
The audio capturer is the fidl manifestation of usb_audio_source.
It's a fidl component manufactured by media_service that exposes a
new interface, MediaCapturer, as well as MediaProducer (the term
'capturer' is awkward, but it's consistent with 'renderer').
MediaCapturer will probably evolve, but the current definition is
adequate for now.

Change-Id: Idfa3388169da15545c31e09374261aad63af8bb2