apps/media: allow media components to survive while connected
Generally speaking, the lifetime of a component vended by media_service
is controlled via the binding that's established when the component is
created. Some components, however, are connected to other components in
graphs and really don't need their primary bindings thereafter. This change
allows decoders, the lpcm reformatter and sinks to survive as long as
there's a binding to the consumer, which means the component is in a
graph.

Change-Id: Ia6be6b9cbe1437c7b11c5cf854fc74b23280232f