apps/media: add FidlConversionPipelineBuilder
ConversionPipelineBuilder has been used for some time to create in-proc
conversion pipelines that convert streams from one type to another. We need
this same capability using fidl components rather then in-proc 'framework'
components. FidlConversionPipelineBuilder uses the same logic as
ConversionPipelineBuilder but instantiates fidl decoders and reformatters
instead of the in-proc versions. It has to function asynchronously, so
it's quite a bit different. This will be used in MediaSource and MediaSink.

Change-Id: Id924b737d38bf14e4c92fe8eab6f5baebaab8c51