apps/media: overhaul MediaSource
Prior to this CL, MediaSource was unused, and its implementation dated back
to when we were OK with building complex in-proc graphs. This CL updates
MediaSource so it enlists the fidl demux, decoders, etc rather than the
in-proc ones. MediaSource needs a test, but it *has* been tested with a
MediaPlayer implementation that uses MediaSource and requires it to do all
the heavy lifting wrt media type conversion.

Also adds some includes to fidl_conversion_pipeline_builder.h that should
have been added in a prior CL.

Change-Id: I0807b688989095c0a1df80832a7b65c6c49bc7f2