apps/media: overhaul MediaSink
Prior to this CL, MediaSink built its own in-proc type conversion pipeline.
When no type conversion was required, it was still in the graph doing nothing
but unnecessary packet copies. This CL makes MediaSink an aggregator for
other fidl components, namely a renderer and any required media type
converters. The sink itself is never in the graph, and, if no conversion is
required, there's no longer any do-nothing component touching the packets.

The audio renderer was updated to allow its media type to be set *after* its
consumer is connected. The new sink wants to do things in this order, and
it makes sure that no packets arrive until the media type is set on the
renderer.

Change-Id: I9eb66b660098c4c11f696c3dbb19cbf300807fed