apps/media: change player to leverage upgraded source and sink
The player has been working overtime due to past shortcomings of the source
and sink. Specifically, it's been dealing with the demux directly and adding
decoders to the graph explicitly. This CL makes the player fully leverage
source and sink capabilities. The player instantiates a source rather than
a demux and allows the upgraded sinks to add decoders as needed.

Change-Id: I44c8fe5e24c11ce31c0128a4d29e767a6dbe3878