apps/media: add MediaPlayerNetProxy to media service
This CL adds a new fidl component that media service will create for clients.
MediaPlayerNetProxy is a proxy for a media player on device machine. The
client asks media service to create one, specifying the device name and the
service name for the remote player. MediaPlayerNetProxy connects to a stub
for the remote player and exposes the MediaPlayer interface so a local client
can control it.

Change-Id: I9b0f16a2597b602c094f8318f39ed4dac43511dc