apps/media: remove layout logic from the video renderer
The video renderer was implementing old layout logic: it would shrink the
video to fit the view horizontally or center the video along the top of
the view if it was smaller than the view, all while maintaining the
original aspect ratio. By doing this, it was preventing higher layers
(like the upcoming flutter widget) from implementing their own layout
logic. This CL just scales the video to fit the view, leaving the rest
up to higher layers.

Change-Id: Icd53835c868f4d3ec662577273f16350c559664c