apps/media: set the timeline on renderers before seeking
When the media player seeks and then primes, renderers encounter packets
with PTSs that start at the new seek position. If the timeline on a
renderer isn't set properly, the renderer may throw away some of those
packets thinking that their presentation time has already passed. This
happens in many cases, for example when we seek position zero when
restarting playback. This CL makes sure the timelines on the renderers
are properly set before packets from the new position can arrive.

Change-Id: I2ce1f9377df84eca6789490a62e7a023c36eb515