apps/media: add media player controller to lib/flutter
This CL adds MediaPlayerController to apps/media/lib/flutter. The controller
manages a MediaPlayer service on behalf of UI code, simplifying the model
and translating some aspects to Dart idioms. A controller can outlive a
MediaPlayer widget so the underlying components can be initialized before
the widget is shown and so that media can continue to play after the widget
disappears (e.g. to play music in the background). The widget itself will
appear in a subsequent CL.

Change-Id: I156336d2d4fd976d1f1eddd51c6639e25785f4fd