apps/media: remove dependencies on src/net
This is the first in a series of CLs that changes the way remote
control of players is done in apps/media. The first step is to
remove dependencies on apps/media/src/net so that some of its files
can be migrated to apps/media/src/net_media_service (a new dir)
without losing history.

The new remote control scheme addresses the problem that the media
player cannot be told by a remote control point to play a new URL.
MediaPlayer has never been allowed to open URLs to a avoid making
it an 'ambient authority' that needs to check client permissions
to determine if a URL can be opened. Instead, a client specifies
the content to play in the form of a SeekingReader (essentially
an already-open URL). A remote control point has no way of doing
this.

The new approach is to create a new service (net_media_service)
that is intended to run in the user's context and so is naturally
subject to the capabilities of the user. net_media_service will
vend two types of agents: NetMediaPlayer and NetMediaPlayerNetProxy.
NetMediaPlayer is pretty much like MediaPlayer (it wraps a
MediaPlayer) except that it takes URLs instead of readers and has
explicit support for publishing itself via NetConnector.
NetMediaPlayerProxy is a proxy for a remote NetMediaPlayer.

Among other things, this change entails moving some files from
src/net to src/net_media_service. This CL removes any dependencies
on src/net, so its files can be moved in a subsequent CL without
breaking the build.

Change-Id: I483e4e24d14680593b6114319bc914e5b1d44e5b