apps/media: allow audio_player to stay around as service
Prior to this CL, the audio player would terminate whenever it
encountered end-of-stream. This is great for simple command line
playback of an audio file, but makes no sense when we want to
control it remotely for e.g. audio flinging. This CL adds the
--stay command line option and makes the player stay by default
if no URL is supplied or the service name is specified, both of
these conditions indicating that the user intends the player be
used as a remotely-controlled service.

Change-Id: Ib5c8d356ebfdb24e95cbbdc2ff5d8546dc7383a5