apps/media: deal with unsuppored formats without crashing
The media code has historically dealt with unsupported formats by
DCHECKing. This does a good job of letting us know we have more work
to do, but it doesn't allow experiences to fail gracefully. As an
example, we're unlikely to support AC3 audio anytime soon (other than
by sending it over HDMI), but we'd like to be able to at least play
video for an A/V file that has AC3. This CL accomplishes a lot of
this by providing a way to represent unsupported encodings and by
failing gracefully when we encounter them. Subsequent CLs will
provide a means of notifying apps when audio and/or video aren't
working due to lack of support.

Change-Id: I22bfd636e26a0194fd39fc8d873922c4f78d4aec