apps/media: handle content with redundant streams
Some content files have more than one stream of a given medium. For
example, movies often have multiple audio streams for different
languages or commentary. Eventually, we need to be smart about
picking a good default stream, and we should allow the application
to enumerate them and choose which to render. In the mean time, this
CL prevents the player trying to render all of them with one sink,
which causes the player to fault.

Change-Id: If81321dad1728bb554c9bcc246447d161971f966