apps/media: handle frame layout changes in video decoder
Aspects of video media types, 'coded size' in particular, can change
through time. The most common case of this is with H264 streams, for
which we get no size information up front. We have to actually start
decoding a frame before we know what size it is (though it generally
doesn't change after that). This will require us to update downstream
components with the new media type information. This CL just deals
with the frame layout implications. Specifically, the video decoder
based its (output) frame layout on the original media type. This is
a problem when the original media type has a coded size of 0,0 and
we try to decode e.g. a 704x480 frame into the buffer allocated based
on that layout. This CL moves the frame layout code into a separate
class (FfmpegVideoFrameLayout) that caches layout information and
updates it based on current media type information. This will prevent
the page faults we currently see with H264. A subsequent CL will
deal with delivering the media type changes downstream so the
renderer can adapt.

Change-Id: Ie6228c3821e7ee5010aed26466aa067f3cb82c45