apps/media: audio renderer ignore prime when disconnected
Renderers that aren't connected need to ignore the Prime request (and
immediately call the callback) if they're not connected. The video
renderer already does this by virtue of not having the logic to wait
for demanded packets to arrive. This CL adds this check to the audio
renderer.

The media player is given renderers that may or may not be relevant
for a given piece of content. Typically, the audio renderer is used
and the video renderer is not (for audio-only content). There are
some cases, however, in which the video renderer is used and the
audio renderer is not. Most often, this is because the audio is of
an unsupported type (e.g. AC3).

The MediaTimelineController is not aware of which renderers are in
use and which are not. It's up to the renderer to ignore timeline
control requests (such as Prime) when not in use.

Change-Id: I0c08d32be0c2a34100547fae8526201352b42738