apps/media: distinguish unsupported container vs not found
The demux, source and media player were reporting an 'asset not
found' problem when the asset existed but we had no support for the
container type (e.g. mp4). This CL addresses the issue by explicitly
checking the reader's 'describe result' before trying to instantiate
an ffmpeg demux. If the describe result is bad, that means the asset
wasn't found. If the describe result is good, and we still can't
create an ffmpeg demux, that means the container isn't supported.

Change-Id: I3d0fbc0191c680fde304fc422a2d91f2f1f63666