apps/media: report stream status to clients
This CL adds information to status reporting on the demux, source
and player indicating the presence of audio and video streams in the
content and whether those streams will be rendered. The media player
examines the stream types reported by the source to determine whether
audio and video are in the content. The demux, source and player all
report whether the demux's producers for audio and video are
connected. A demux producer won't be connected if:

1) The content doesn't contain a stream with that medium, or
2) The client didn't provide a renderer for that medium, or
3) We're not done hooking things up, or
4) We don't support the media type of the stream with that medium.

The player also synthesizes a Problem structure to attach to its
status if no streams could be rendered. If audio or video can be
rendered but the other medium cannot (due to 4), the client can
detect this itself based on the status information provided.

Change-Id: Id769bf572887c89dc27e4bfeaa3849a3a38a6a10