apps/media: add resampler to src/audio
This CL adds an audio resampler, which will eventually be used to
build a new system mixer. The resampler class is parameterized on
the sample type and the type used for subframe indexing (i.e. the
type used to tell us where we are between two input frames). For
the moment, only the <float, float> specialization is complete.
Channel count is a constructor parameter, but at some point, we'll
want to write stereo-specific code so the channel loop can be
unrolled for that common case. The resampler handles arbitrary
input and output buffer boundaries.

Change-Id: I4f1026b533fd62b43069c22482133212308c6fe6