apps/media: reader shouldn't DCHECK on ERR_PEER_CLOSED
When you try to play a URL such as www.youtube.com, the fidl reader
DCHECKs because the code doesn't recognize the status that's returned
from the read. This CL fixes that, recognizing the specific status
that's produced in this scenario and using FTL_LOG(ERROR) instead
of FTL_DCHECK for unrecognized status values.

Change-Id: I6e3a41ae8452e27aa5b5325d076bbe134553900a