apps/media: complete prime in video renderer when priming is done
There's been a TODO in the video renderer for awhile to postpone
completing the 'prime' method until priming is actually complete.
This CL addresses that. Priming is complete when the initial packet
demand is met. This fix addresses a possible race in which the
video renderer is told to start playing before it has the packets
it needs to do so without starving.

Change-Id: Ica2e55e518df82ea901604b1509084a8c664dac1